requirement.txt
python-telegram-bot==21.6
openai==2.6.0
python-dotenv==1.2.1